raid1: add raid1 specific attributes to extent
This patch adds the setting of raid1 specific attributes to each extent,
and their retrieval during an import. This is to prepare the remove of
these attributes from file names.

Also removes the oid from the xattrs of the entries, as the information
is already in the filename. Change the tests aswell following that
modification.

Change-Id: I52915ef1755dee121d4f1a78f401a8d6f0b5f6fe
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12732
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>