ldm, tlc: rename "reload" to "refresh" and add it to ldm
TLC reload operation and option deal with the tape library status
cache and not with a tape that one wants to unload and load again
into a drive. This patch replaces "reload" by "refresh" to avoid
any ambiguity between tape movement and tape library status cache
update.

This patch adds the refresh action to ldm API.

We replace in admin API the direct request to tlc to refresh it by
the new ldm call.

Change-Id: Ie88631ac0c18dc0f808b680bc7b8c4fc170982eb
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12766
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>