layout: change file name at creation and read
This patch changes the way the file name is created for both put and
get. It is changed from "oid.version.layout_tag.object_uuid" to
"oid.extent_uuid". The goal is to make this name more generic and easier
to read by different layouts without removing its unicity.

For this, the extent UUID must be available when creating an extent,
meaning we can't wait for the extent to be added to the DSS. Thus, we
now create the extent UUID in the layout, by using the libuuid library.

Change-Id: Ief856727383e7c67adf590c21d8fab6bac4b1fe7
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12784
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>