import: split attr retrieval in two
This patch splits the retrieval of the xattrs in two. The first part
will be done by the io-module as part of a "general" xattrs retrieval,
which includes the uuids, the user md, ... The other part is done by the
layout, and will retrieve layout-specific attributes. For raid1, this
includes the replica count, for raid5 it may include the type of raid5
used or the parity check, ...

Change-Id: I1e5ed367d75ee2fc3406a9a08f4f804f3f4faa28
Signed-off-by: Yoann Valeri <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12788
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>