lrs: use dss_medium_health and dss_device_health
The health is necessary when adding a medium into the cache. The I/O
scheduler for write allocation request also fetches a list of media from
the DSS. Since this selection algorithm will use the health as a
criteria to choose the best medium for an allocation, it also needs the
health at this point.

For the device health, there are two cases:
- at startup when we load all the usable devices from the DSS;
- when we add a device through the notify request.

Change-Id: I707dbc43fa2ff979c0e59b35dbb60fdf021db325
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12906
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>