ldm: remove log parameter from ldm_lib_open
The scsi ldm lib open is now only a connection to the tlc that is
generating no log. The rados ldm lib open was not using the log
parameter and there is no dummy lib open at all.

There is no reason anymore to maintain a useless log parameter
in the ldm lib open call.

Change-Id: Ib9bebe6ddc3be8589152aa5e3d26823058b2c6de
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12793
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>