import: improve the logic of inserting found objects
This patch improves the procedure to find where to put the objects
after an import, following the documentation.

More specifically, if an object already exists with the same oid as a
file found after an import, the latter is added to the DSS with the
suffix ".import-<current date>".

Also add a corresponding test.

Change-Id: I3865131214f43cd8fcf956faa5734ab3636cbad9
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12822
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>