bugfix: fix invalid rc retrieval in test driver
This patch fixes an issue where the return code of an executed test in
the test-driver would be retrieved after an echo instead of after the
test. This means the return code would always be 0, and never be the
return code of the actual test, meaning every test would be marked as
passing successfully.

Change-Id: I654a075cf8d4c82dd44c54a355a1e01fe467f575
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12877
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>