admin: remove device from db if device_add fails
This patch adds the removal of devices from the database if the
device_add function fails after inserting them. For instance, this can
occur if one tries to add a drive with the LRS started but not the TLC.
In this case the notify will fail, so the device will be removed from
the DB aswell, ensuring it can be added again if the TLC is restarted
afterwards.

Change-Id: I0d0b10a0484829e07bcb7d87feff116a33d554ac
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12896
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>