lrs: fix memory leak on release no sync
The request container of a release request is freed by the device thread
when the sync is finished. But a release that do not require
a sync will not be sent to the device thread. The thread handling the
release must free the request container of a "no sync" release.

Change-Id: I325c45749f954fba44aeed56fc5a91ab70db74a4
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12924
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>