lrs: remove device from io_sched when removed from the LRS
A device can be removed in two contexts:
- when the LRS stops, all the devices are removed;
- when a device is locked is is removed through the notify request.

In both cases, we need to remove the device from the I/O scheduler as
well. This introduces a dependency between the lrs_dev handle and the
I/O scheduler. Therefore, the test test_lrs_device.c needs to setup an
I/O scheduler.

Change-Id: I5406f4da022e42b37c45c9693abd67ef9bb008f1
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12926
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>