refacto lrs: extract allocation result handling from read alloc
The function sched_read_alloc_one_medium has become very complex to
understand as it manages both the first allocations and the error
handling after an allocation failure from a device.

To make the code simpler to understand and modify, both of these use
cases will be moved to dedicated functions. The first step is to extract
the common parts of both cases into a function sched_try_alloc_medium.

Change-Id: I58db79f1b1b69aa50719356b2ef85991ce2d7cd9
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12983
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>