lrs: check compatibility only for tape
The function compatible_drive_exists called when no device can be found
for a medium in the scheduler was calling tape_drive_compat for all the
types of resources. This leads to the LRS thinking that a dir medium is
not compatible with its device. In some cases, the LRS would return
ENODEV to the client.

Change-Id: I13d31c35d4299322dc67f4852c46aaacfcdcff27
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12981
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>