lrs: check write permission in dev_picker
In find_write_device, dev_picker is called 2 times without a medium and
is_write is true. The purpose of these two calls is to find a device
whose loaded/mounted medium can accept the new object we want to write.

Then, dev_picker is called a third time if no device was found. This
time a medium is provided and is_write is false. This last call wants to
find a device that is free for a new allocation. This device may contain
a medium that was not suitable for I/O since the two first calls to
dev_picker returned NULL at this point.

The current implementation has two issues:
1. since is_write is set to false, we never check the write permissions
   and tags on the medium used for the I/O;
2. even if is_write was true, dev_picker only checks the permissions of
   the loaded/mounted medium, not the permissions of pmedia. Which means
   that dev_picker can choose a medium that is failed for example.

Change-Id: I9d538e0b3b54c8f02a917d168d5034b4af7a14be
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/12982
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>