lrs: do not manage EBUSY error when moving a tape
This error was used before the introduction of the TLC to indicate that
a drive was temporarily unavailable during a move. In this case, the LRS
would retry to move the tape. This behavior was removed with the
introduction of the TLC. This patch removes the checks for EBUSY that
where done in this case. If we want to reintroduce this behavior, it
makes more sense to move it to the TLC as this is a behavior specific to
tapes and is related to the low level behavior of the tape library.

Change-Id: I93880fc3cb9f39efb5a68f615f2654a159b8322b
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13027
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>