lrs: do not unlock the device's medium twice
A medium associated to a device is freed when its health reaches 0. If
at the same time, the device's health reaches 0, the device thread will
stop. When stopping the device thread we must make sure that we don't
unlock the medium again. The second unlock will fail since there is no
lock.

This patch solves the issue by cleaning the lock after the first unlock.
Then the device thread checks whether the lock was set or not before
unlocking the medium.

Change-Id: I40fe3fabad9ee0667be8d857d9da7057fb0d881b
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13032
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>