cfg: remove useless warning of PHO_CFG_GET_INT
When an integer value is not found in the configuration, we
return the failover value and print a warning message. This
warning message is useles when we are simply checking if a value
is defined in the configuration or if we need to use the default
value.

This patch moves the message to the debug level.

Change-Id: Ib17f98f50aae18f5277de9fbdedc7445e72bf94e
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13079
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>