comm: guard freeaddrinfo() calls
This fixes a couple of problems with freeaddrinfo():
- pho_comm_open() would call freegetaddrinfo on NULL in the unix server
case or on error in tcp_set_socket_addr_path(); this works on glibc
but is not guaranted by posix and would crash on e.g. musl
- calls freeaddrinfo twice in pho_comm_open() if listen(),
epoll_create() or epoll_ctl() fail. This is UAF even on glibc.

guard all freeaddrinfo() calls with a NULL check, and set the address to
NULL after free when there is a chance that it could be used again.

Change-Id: I9ba9aa9028a4c97b3e4258f083daad2d71042f1b
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Signed-off-by: Dominique Martinet <asmadeus@codewreck.org>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13161
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>