podman: add a containerfile and postgres script
This patch adds a simple containerfile containing all the repo and
commands to start a Podman container. It also adds a simple postgres
script used to properly initialize the postgres service. This script
should be used after the container is started.

Also modify some tests to comply with container limitations, mostly due
to `losetup` being unusable.

Change-Id: I8533bbfe025c8d9fafe9d012cb6aa7f4bafe763d
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13122
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Arnaud GAUTHERET <arnaud.gautheret@cea.fr>
Reviewed-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>