raid4: rename raid5 to raid4
The advantage of raid5 over raid4 is in the distribution of the parity
across disks. This has a performance advantage for disks as it avoids
write contention when updating blocks. Since phobos is an object store,
there is no update of parts of an object (at least at this moment). So
it makes more sense to have a RAID4 layout. RAID5 can be added later if
needed for disk based families.

Change-Id: I763ee49b0cd981f9ffe5b2f03e0c3702dcc384d8
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13240
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Gauthier Evraerd <gauthier.evraerd.ocre@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>