io-sched fifo: do not increment medium index on retry
In the FIFO I/O scheduler, the field queue_element::num_media_allocated
represents the number of elements actually allocated to an allocation.
When a medium is returned by generic_get_device_medium_pair,
num_media_allocated is increased to reflect this. When the caller
retries an allocation (if the medium is temporarily unavailable for
example), we must decrement num_media_allocated to reflect the fact that
the medium was not actually allocated in the end.

Change-Id: I0af8231588d5f0621b528484c975b7949135890f
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13239
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Gauthier Evraerd <gauthier.evraerd.ocre@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>