io-modules: add pho_iod_from_fd to the I/O descriptor API
This function will be implemented as a callback of the I/O decriptor
implementation. The goal is to convert an openned file descriptor to an
I/O descriptor. This will allow Phobos to reuse the read and write
functions that are currently reimplemented in the layout-modules.

The layout can convert the file descriptor of the Xfer to a pho_io_descr
struct that can then be used to perform POSIX I/O.

Change-Id: I07bc53fca88899532c01530c68585937214406a4
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13300
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>