io: add io_adapter_module in pho_io_descr
To use a pho_io_descr, we need the corresponding io_adapter_module.
It is the io_adapter_module that contains the
pho_io_adapter_module_ops struct to access the function pointers of the
module. To avoid having to keep track of both the io_adapter_module and
the pho_io_descr, this patch adds it to the pho_io_descr. Making it a
self describing structure.

Change-Id: If88f4be2d5a370f1de41f65ce86a1a14df6a30a9
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13245
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Gauthier Evraerd <gauthier.evraerd.ocre@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>