raid4: split raid4 module in several files
This patch rewrites the RAID4 implementation to make it simpler.
Most of the actions done for the RAID4 are also done in the RAID1 and
can be mutualized in raid_common.c.

struct raid_ops now contains read_split and write_split callbacks.
These callbacks are the only part that is specific to the actual RAID
implementation. They are called with opened pho_io_descr for each
media selected by the LRS. All the implementation has to do is write or
read data appropriately and set attributes specific to the layout on the
medium.

Change-Id: I775dded79b0bfd024f5efe225088e99cd94be68e
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13246
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>