lrs: avoid already allocated drive selection
In case a client try to allocate a second drive but there is only one
available, the scheduler will consider the first drive as selectable and
wait for its release, which provokes a deadlock.

Now, a drive allocated by the same client cannot be considered
selectable if both are needed for the same operation.

A boolean is also added to the communication protocol with the LRS to
distinguish client which requests multiple allocations for one operation
(repack) or for multiple operations (mput). The boolean
`prevent_duplicate` will be set to `true` if allocations are bound to
the same operation and `false` otherwise.

Change-Id: Ife728cefea7ceac6c46f0669179a12ad43ff812a
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13235
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>