cli: properly format JSON output for list commands
In order for json.dumps to properly format the output, the value passed
to it must not be pretransformed. dumps will properly convert nested
lists and dictionnaries to JSON values. This patch adds a new parameter
to get_display_dict to specify the default conversion function which was
str. For the JSON format, we can now specify it to be the identity
function (i.e. do not transform the value).

Change-Id: I358434f95e7151812465da3989f568473fbed9ad
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13268
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Gauthier Evraerd <gauthier.evraerd.ocre@cea.fr>