raid4: get block size from extent on get
Add a callback get_block_size to ask the layout for the block size when
reading an extent. This was implemented as a callback since the block size
is computed before the buffers are allocated. To avoid allocating the
buffers in the read_split callback for each RAID layout, we simply ask
the layout for its block size before the allocation.

For RAID4, the block size to use is the chunk size of the layout used
when writing the data.

Change-Id: I6ab145e18444071798c258b8130722f914a41ac5
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13299
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Gauthier Evraerd <gauthier.evraerd.ocre@cea.fr>