dss: fix GC invocation for a repack operation
The previous GC request selected all deprecated objects, because the
request did not consider if the object had an extent on the targeted
tape, only if such an object existed. This is now changed.

A test is also added to check deprecated objects from other tapes are
not removed from the database.

Change-Id: Ie1747cbbeb6b21e5eb66acc95ddb2afa9939c8e7
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13290
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>