proto/lrs: add nb_extents_written to release request
Until now, the number of extents written per release request was assumed
to be 1. With the repack operation, the write allocation may be
requested for multiple extents. Now, this number is sent to the LRS and
tape stats are updated accordingly.

A check for the tape stats during the repack operation is added in
tests.

Change-Id: Ibf250066ac1dfada212acca986ca5ee5d470a99e
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13291
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>