dss: add options --sort and --rsort with drive list
This patch adds a new option to sort the output of a "phobos drive
list". The sort can be ascending (--sort) or descending (--rsort) and
it is done at the postgresql level.

Also, a new structure, dss_sort, has been added. It contains the
information necessary to sort.

Finally, this patch adds tests for this option.

This option will be added to the media, object and extent in the next
patches.

Change-Id: I1c553fa4c62dbb7bbbab45a3d0e4820564bfc144
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13292
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>