dss: add options --sort and --rsort with extent list
This patch adds the new options --sort and --rsort with "phobos extent
list". Only the layout, size, extent_count, oid, object_uuid and
version are supported by the sort. The others attributes are not
supported because sorting is not possible in psql for them. The select
query returns a list of the attribute for each extent, if there are
several extents for an object. The sort of these attributes becomes
specific to each attribute.

The only exception is for the "size" attribute, the sort is done at
the C API level. The sort is done by comparing the sum of the size of
all the extents of an object.

Change-Id: I9ddff9bae8ac041f1158f18618f91d9d53cf0d91
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13342
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>