test_lrs: cleanup traps in the subtests
test_lrs is subdivided in a lot of subtests that are independent of each
other. A complete setup/cleanup is done for each sub test. This is done
using traps inside each subtest to ensure that everything is cleaned up
properly.

The issue is that our custom test-driver does not work well with traps
inside the subtests. For example, drain_all_drives calls mtx unload
which displays logs when unloading a cartridge. These logs are displayed
on the stdout of test-driver and not in the log file.

To avoid this, the cleanup of tests should be done outside of traps. So
this patch calls cleanup functions explicitly at the end of each
subtest, removes the unnecessary traps and remove the cleanup function
from all the traps because it is already set as the global trap.

Change-Id: I65ddf6a38d45a1fe9ffe514a97a9e9bb7aaaab36
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13376
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>