raid4: check hash of extents with phobos get
This patch adds the hash of the extents when reading and compares
them with the hashes obtained when writing, to check that the extents
are not corrupted. If the hashes are different, an error is returned.

Also, the hash are calculated based on the hash contained in the
extent. If we disable or change the hash algorithm, the hash will be
calculated in any case with the hash algorithm used at the writing.

Change-Id: I186ee44beb45b0b8e3ccec2009608c30bd5c934a
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13364
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>