raid4: make test_raid4 generic enough to be used for raid1
To prepare for the mutualization of RAID1 with the new raid_io_context,
this patch renames test_raid4.test into raid_layout_common_tests.sh.
The type of raid is extracted from the RAID_LAYOUT in the test setup.
This will allow the tests to be run for RAID1 as well.

Change-Id: I78121aeada7ad555a505842ee80ef3611f0ea760
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13372
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Gauthier Evraerd <gauthier.evraerd.ocre@cea.fr>