test raid1: run raid_layout_common_tests with RAID1
The test is adjusted to match the different data layout. More
specifically, the tests with a split need to be run with a file slightly
smaller than the total size on dir. There are 6 dirs of 1M which is 6M.
The RAID1 test uses a repl_count of 2 with a file of 3M.

Since this does not fit, the files are made slightly smaller with 2740KB
of data.

Change-Id: I8d91aae0adb10cb191658f9655ea53c2a0b10398
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13373
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>