cli: add dir delete
This patch adds the new command "phobos dir delete". It takes as
argument a list of dirs to remove. Since dirs are a special case, we
first delete the device associated with the dir and then the media. If
the media removal fails, the device that was already removed is put
back.

Change-Id: Iff942563c3963db496b70b631ab7c51cb6994dcd
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13387
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Gauthier Evraerd <gauthier.evraerd@cea.fr>