proto: add operation to read allocation request
A new field is added to the read allocation request protocol to allow
multiple kind of operations to be managed by the scheduler. It will be
useful for the HSM mechanism and the deletion of objects, to also
consider delete requests.

Both read and delete requests are consider as one kind of requests by
the scheduler as they both directly targets media.

Change-Id: I0b7f8a9c3bebc9658c59d72cd9e6792526e746fc
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13673
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Gauthier Evraerd <gauthier.evraerd@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>