raid1: add the possibility to check extent hash on get
This patch adds the possibility to compute the hash of extents in the
RAID1 layout. To make this possible, we need to have access to the
buffers in userspace. Therefore, get will use a standard read/write loop
using ioa_read and ioa_write to have access to the buffers.

Change-Id: I81ec004dffbcffd914bb42b33096658fb783bb5c
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13423
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>