test: fix extent corruption method in extern cli raid common test
To corrupt an extent, the previous version of the test was
writting the 0xFF byte at the beginning of the extent. But the
extent was produced from /dev/urandom and could contain 0xFF as
first byte. This corruption method is used so many time into this
test that the test was often failing.

The new corruption method is adding 1 to the first byte to change
its value.

Change-Id: I8901b7d4bd2775770d0f5b5ab863b586ed7b25dc
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13572
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>