store: add hard delete behavior to `phobos_delete` call
This behavior will allow one to remove from the DSS information of the
deleted object and its layout. The extents information are still
maintained in the database, but their state is set to orphan.

This is the behavior stated by the design document for tapes, as we want
to keep stats of written data as long as the tape is not repacked.

Tests will be implemented in a future patch.

Directory (and RADOS) behavior will be implemented in a future patch.
Until now, they will both act as tapes.

Change-Id: I99f20a255689afe2e7ad1dd7b99ff394780e96f2
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13774
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Gauthier Evraerd <gauthier.evraerd@cea.fr>