admin: add phobos_admin_drive_scsi_release
This patch adds a new phobos admin function: drive scsi_release. It
can be used to release the scsi reservation of one or many drives. It
is to be used in case phobosd crashed while a tape was in use
(read/write).

This function only takes care of the scsi reservation for the moment.
The management of locks, status of tapes and drives, unmount the tape
are left to the admins.

Change-Id: I19e478516bd4495a8879251dc3ecadf5782badb1
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13756
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Gauthier Evraerd <gauthier.evraerd@cea.fr>