cli: add phobos drive scsi_release
This patch adds the command phobos drive scsi_release in the cli.
It takes as arguments one or many drive(s) to release the scsi
reservation.

This command should be used with caution. You have to be careful with
the DSS locks, status of the tapes and drives, is the tape still
mounted?

Change-Id: I291892e28d92121cf80fc2b8877cf2eda9f86a9b
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.ocre@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13757
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Gauthier Evraerd <gauthier.evraerd@cea.fr>