admin: add path normalization in phobos commands
This patch allow one to use a relative path will phobos commands. This
patch add a new common function to normalize a path. Path
normalization is done in the API, except for the put/get and dir list
where the path is used and replaced by a file descriptor or a filter
in the cli.

Also, as some functions for devices accept the serial number instead
of the path if it is a drive, the normalization is only done if it is
a dir.

Change-Id: Ic1fd20d66d894b7eebbb3cf5222026e2416746ad
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13788
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>