cli: add new option --new-library in phobos drive migrate
This patch adds the new option --new-library to change the library of
a drive to phobos drive migrate. It can be used with the other options.

Also, host has switched from positional option to optional option.
phobos drive migrate can be used with either --host or --new-library
or with both. At least one option should be specified.

Change-Id: Ic158d9bd4ba159a5fe8f829745f5e5f3aa838a3b
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13830
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>