dss: add source resource list to dss_generic_update
The function dss_generic_update takes now a source and destination
resource list as argument. For a simple update (not a primary key),
source and destination are the same resource list. Otherwise, to
update a primary key, source should contains the old value and
destination the new value of the primary key.

Change-Id: If9ee8a8d8443c173b39eb9d3a227986e330dc78c
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13835
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>