proto & lrs: add no-split option to proto
This patch adds the no-split option to the proto of write request. If
this option is true, all the data of the request should be written on
the same medium.

Also, the LRS handle now the option no-split. If there is no medium of
the size required, the LRS return an error to the client.

Change-Id: I6bb3751d0762abf4b222207fad27d3a93dae1b89
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13915
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>