proto & lrs: add a boolean to the release request to do only a sync
This patch adds a new boolean "partial" to the release request to know
if we need to do only a sync of the medium or a normal release. The
value 'ld_needs_sync' is set to true when receiving a partial release
request to avoid waiting until the threshold are exceeded.

Change-Id: I59bf38699a3ec6641deef1f2096df4143366be30
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13939
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>