xfer: modify xfer structure to hold N objects
This patch modifies the xfer structure to now contains N
objects during a transfer. For this, the pho_xfer_target structure is
added. This structure contains the information of an object. A list of
targets representing all the objects has been added to the pho_xfer_desc
structure, as well as the number of objects (target) present.
In addition, the size that was present in the put params has been
moved to the new pho_xfer_target structure.

Finally, for now xfers only contain one object, the management of N
objects in xfers will be done in the next patches.

Change-Id: Iacb0a94de014a56f3749d81fabaa49f739608896
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13987
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>